chore: Remove Forest Watcher infrastructure
It looks like `apigw.tf` shouldn't be a part of this repo. there have been discussions:

Excerpts from Slack...
One thing I’m not sure about. The PR (which is from staging) contains terraform/apigw.tf.
However, it looks like this shouldn’t be added based on the comments of an earlier PR to get terraform/apigw.tf into dev:
https://github.com/wri/gfw-aws-core-infrastructure/pull/81#pullrequestreview-945862640

In staging:
1. It’s added: https://github.com/wri/gfw-aws-core-infrastructure/commit/a876ab1483a7a69364df8448d5f65c3d56ca4521
2. It goes away: https://github.com/wri/gfw-aws-core-infrastructure/tree/c0e0888be613c8e0852faa4e6ab35e549f216f96
3. It shows up again: https://github.com/wri/gfw-aws-core-infrastructure/tree/3cd5656ed67dc06b723a614378e389406d48c1b6
4. It disappears again :magic_wand: : https://github.com/wri/gfw-aws-core-infrastructure/tree/e4d2ac25e68630d0b32cc1cd4283b187def2451c
And then there’s the PR (above) to add it again. :shrug:
It looks like it doesn’t belong in the repo and should be removed.

GTC-1921